+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".WelcomeActivity">
+        <activity android:name=".Controller.WelcomeActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
@@ -21,11 +21,13 @@
                 android:name="preloaded_fonts"
                 android:resource="@array/preloaded_fonts"/>
 
-        <activity android:name=".LeagueActivity">
+        <activity android:name=".Controller.LeagueActivity">
         </activity>
-        <activity android:name=".BaseActivity">
+        <activity android:name=".Controller.BaseActivity">
         </activity>
-        <activity android:name=".SkillActivity">
+        <activity android:name=".Controller.SkillActivity">
+        </activity>
+        <activity android:name=".Controller.FinishActivity">
         </activity>
     </application>
