--- app/src/main/java/paulo/com/swoosh/Constants.kt
@@ -1,3 +1,0 @@
-package paulo.com.swoosh
-
-const val EXTRA_LEAGUE = "league"
