+++ app/src/main/java/paulo/com/swoosh/Controller/SkillActivity.kt
@@ -1,0 +1,44 @@
+package paulo.com.swoosh.Controller
+
+import android.content.Intent
+import android.os.Bundle
+import android.view.View
+import android.widget.Toast
+import kotlinx.android.synthetic.main.activity_skill.*
+import paulo.com.swoosh.Utilities.EXTRA_LEAGUE
+import paulo.com.swoosh.R
+import paulo.com.swoosh.Utilities.EXTRA_SKILL
+
+class SkillActivity : BaseActivity() {
+
+    var league = ""
+    var skill = ""
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_skill)
+        league = intent.getStringExtra(EXTRA_LEAGUE)
+        println(league)
+    }
+
+    fun onBallerClick(view: View) {
+        skill = "baller"
+        beginnerSkillBtn.isChecked = false
+    }
+
+    fun onBeginnerClick(view: View) {
+        skill = "begginer"
+        ballerSkillBtn.isChecked = false
+    }
+
+    fun onSkillFinish(view: View) {
+        if (skill != "") {
+            val finishActivity = Intent(this, FinishActivity::class.java)
+            finishActivity.putExtra(EXTRA_LEAGUE, league)
+            finishActivity.putExtra(EXTRA_SKILL, skill)
+            startActivity(finishActivity)
+        } else {
+            Toast.makeText(this, "Please select a skill level", Toast.LENGTH_SHORT).show()
+        }
+    }
+}
