--- app/src/main/java/paulo/com/swoosh/SkillActivity.kt
@@ -1,15 +1,0 @@
-package paulo.com.swoosh
-
-import android.os.Bundle
-
-class SkillActivity : BaseActivity() {
-
-    var league = ""
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_skill)
-        league = intent.getStringExtra(EXTRA_LEAGUE)
-        println(league)
-    }
-}
