+++ app/src/main/java/paulo/com/swoosh/Utilities/Constants.kt
@@ -1,0 +1,4 @@
+package paulo.com.swoosh.Utilities
+
+const val EXTRA_LEAGUE = "league"
+const val EXTRA_SKILL = "skill"
