+++ app/src/main/res/layout/activity_skill.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".SkillActivity">
+        tools:context=".Controller.SkillActivity">
 
     <ImageView
             android:layout_width="0dp"
@@ -46,24 +46,27 @@
         <ToggleButton
                 android:text="@string/beginner"
                 android:layout_width="wrap_content"
-                android:layout_height="wrap_content" android:id="@+id/toggleButton4" android:layout_weight="1"
+                android:layout_height="wrap_content" android:id="@+id/beginnerSkillBtn" android:layout_weight="1"
                 android:background="@drawable/swoosh_toggle" android:textColor="@android:color/white"
                 android:textOff="@string/beginner" android:textOn="@string/beginner"
-                android:layout_marginEnd="16dp"/>
+                android:layout_marginEnd="16dp" android:onClick="onBeginnerClick"/>
         <ToggleButton
                 android:text="@string/baller"
                 android:layout_width="wrap_content"
-                android:layout_height="wrap_content" android:id="@+id/toggleButton5" android:layout_weight="1"
+                android:layout_height="wrap_content" android:id="@+id/ballerSkillBtn" android:layout_weight="1"
                 android:background="@drawable/swoosh_toggle" android:textColor="@android:color/white"
-                android:textOff="@string/baller" android:textOn="@string/baller" android:layout_marginStart="16dp"/>
+                android:textOff="@string/baller" android:textOn="@string/baller" android:layout_marginStart="16dp"
+                android:onClick="onBallerClick"/>
     </LinearLayout>
     <Button
             android:text="@string/finish"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:id="@+id/button" app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="16dp"
+            android:id="@+id/skillFinishBtn" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="16dp"
             app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="16dp" android:layout_marginBottom="16dp"
             app:layout_constraintBottom_toBottomOf="@+id/imageView4" android:fontFamily="@font/montserrat"
-            android:textColor="@android:color/white" android:background="@drawable/swoosh_button"/>
+            android:textColor="@android:color/white" android:background="@drawable/swoosh_button"
+            android:onClick="onSkillFinish"/>
 
 </android.support.constraint.ConstraintLayout>
