+++ app/src/main/res/layout/activity_welcome.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".WelcomeActivity">
+        tools:context=".Controller.WelcomeActivity">
 
     <ImageView
             android:layout_width="0dp"
